Update test project setup in Create-Day.ps1 script
Modified the `Create-Day.ps1` script to enhance the setup of the test project folder. Removed `GlobalUsings.cs` and `UnitTest1.cs` from the `$TestProjectFolder`. Updated the `$TestProjectFile` to use `MSTest.Sdk` instead of `Microsoft.NET.Sdk`. Removed the `<PropertyGroup>` section and added a new `<ItemGroup>` with project references to the main project and shared test project.